Move logging into a separate directory - Use new property to set log level.
