+++ src/logging/logger.ts
@@ -2,10 +2,11 @@ import {getLogger, LogRecord, setup} from "log";
 import {ConsoleHandler} from "log/console-handler";
 
 import formatters from "./formatters.ts";
+import config from "../config.ts";
 
 setup({
     handlers: {
-        console: new ConsoleHandler("DEBUG", {
+        console: new ConsoleHandler(config.logLevel, {
             formatter: (record: LogRecord) => {
                 return `[${record.levelName}] ${record.msg} ${record.args.map(formatters.format).join('\n')}`
                     .replaceAll('\n', `\n[${record.levelName}] `);
@@ -14,7 +15,7 @@ setup({
     },
     loggers: {
         default: {
-            level: "DEBUG",
+            level: config.logLevel,
             handlers: ["console"]
         }
     }
